cronML/requirements.txt
asttokens==3.0.0
attrs==25.3.0
backcall==0.2.0
beautifulsoup4==4.13.3
bleach==6.2.0
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
decorator==5.2.1
defusedxml==0.7.1
docopt==0.6.2
executing==2.2.0
fastjsonschema==2.21.1
Flask==3.0.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.3
idna==3.8
imbalanced-learn==0.13.0
imblearn==0.0
iniconfig==2.0.0
ipython==8.12.3
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mistune==3.1.3
mysql==0.0.3
mysql-connector-python==9.0.0
mysqlclient==2.2.4
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
numpy==2.1.1
packaging==24.1
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
pickleshare==0.7.5
pipreqs==0.5.0
platformdirs==4.3.7
pluggy==1.5.0
prompt_toolkit==3.0.50
pure_eval==0.2.3
Pygments==2.19.1
PyMySQL==1.1.1
pyodbc==5.1.0
pytest==8.3.2
python-dateutil==2.9.0.post0
pytz==2024.1
pyzmq==26.4.0
referencing==0.36.2
requests==2.32.3
rpds-py==0.24.0
scikit-learn==1.6.1
scipy==1.15.2
six==1.16.0
sklearn-compat==0.1.3
soupsieve==2.6
SQLAlchemy==2.0.34
stack-data==0.6.3
threadpoolctl==3.6.0
tinycss2==1.4.0
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.1
ua-parser==0.18.0
urllib3==2.2.2
user-agents==2.2.0
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.0.4
WTForms==3.1.2
yarg==0.1.9
